# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/tests": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves all speed test records with pagination support",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Get all speed tests",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Number of records to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of records to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SpeedTest"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new speed test record and returns the test ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Create a new speed test",
                "parameters": [
                    {
                        "description": "Speed test data",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SpeedTest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/tests/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a specific speed test record by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Get speed test by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Speed test ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SpeedTest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Returns server timestamp in nanoseconds for latency calculation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Speed Test"
                ],
                "summary": "Ping endpoint for latency measurement",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.PingResponse"
                        }
                    }
                }
            }
        },
        "/result/{id}": {
            "get": {
                "description": "Retrieves a speed test result in Ookla speedtest.net compatible format",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get speed test in Ookla format",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Speed test ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OoklaCompatibleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.OoklaCompatibleResponse": {
            "type": "object",
            "properties": {
                "download": {
                    "$ref": "#/definitions/models.OoklaSpeedResult"
                },
                "interface": {
                    "$ref": "#/definitions/models.OoklaInterfaceInfo"
                },
                "ping": {
                    "$ref": "#/definitions/models.OoklaPingResult"
                },
                "result": {
                    "$ref": "#/definitions/models.OoklaResultInfo"
                },
                "server": {
                    "$ref": "#/definitions/models.OoklaServerInfo"
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "upload": {
                    "$ref": "#/definitions/models.OoklaSpeedResult"
                }
            }
        },
        "models.OoklaInterfaceInfo": {
            "type": "object",
            "properties": {
                "externalIp": {
                    "type": "string"
                },
                "internalIp": {
                    "type": "string"
                },
                "isVpn": {
                    "type": "boolean"
                },
                "macAddr": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.OoklaPingResult": {
            "type": "object",
            "properties": {
                "high": {
                    "type": "number"
                },
                "jitter": {
                    "type": "number"
                },
                "latency": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                }
            }
        },
        "models.OoklaResultInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.OoklaServerInfo": {
            "type": "object",
            "properties": {
                "cc": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latency": {
                    "type": "number"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "sponsor": {
                    "type": "string"
                }
            }
        },
        "models.OoklaSpeedResult": {
            "type": "object",
            "properties": {
                "bandwidth": {
                    "description": "bits per second",
                    "type": "integer"
                },
                "bytes": {
                    "type": "integer"
                },
                "elapsed": {
                    "description": "milliseconds",
                    "type": "integer"
                },
                "latency": {
                    "type": "number"
                }
            }
        },
        "models.SpeedTest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "client_ip": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "download_size_bytes": {
                    "type": "integer"
                },
                "download_speed_mbps": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "isp": {
                    "type": "string"
                },
                "jitter_ms": {
                    "type": "number"
                },
                "ping_latency_ms": {
                    "type": "number"
                },
                "region": {
                    "type": "string"
                },
                "server_city": {
                    "type": "string"
                },
                "server_country": {
                    "type": "string"
                },
                "server_name": {
                    "type": "string"
                },
                "sponsor": {
                    "type": "string"
                },
                "test_duration_seconds": {
                    "type": "number"
                },
                "test_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "upload_size_bytes": {
                    "type": "integer"
                },
                "upload_speed_mbps": {
                    "type": "number"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
        "types.PingResponse": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "description": "Server timestamp in nanoseconds",
                    "type": "integer"
                }
            }
        }
    }
}